src/ml/style_analyzer.py
"""
Style analysis and matching using computer vision.
"""
import numpy as np
from PIL import Image
from typing import Dict

class StyleAnalyzer:
    """
    Analyze and classify interior design styles using ML.
    """
    
    def __init__(self, model_path: str = None):
        """
        Initialize the style analyzer.
        
        Args:
            model_path: Path to the trained model
        """
        self.model = None
        self.model_path = model_path
        self.styles = [
            'modern', 'contemporary', 'minimalist', 'industrial',
            'scandinavian', 'bohemian', 'traditional', 'rustic'
        ]
    
    def analyze_image(self, image_path: str) -> Dict[str, float]:
        """
        Analyze an image and return style probabilities.
        
        Args:
            image_path: Path to the image file
        
        Returns:
            Dictionary with style names and confidence scores
        """
        # TODO: Implement image analysis with CNN
        style_scores = {}
        
        return style_scores
    
    def match_preferences(
        self,
        user_preferences: Dict,
        available_items: list
    ) -> list:
        """
        Match user preferences with available items.
        
        Args:
            user_preferences: User's style preferences
            available_items: List of available furniture/decor items
        
        Returns:
            Sorted list of matching items
        """
        # TODO: Implement preference matching algorithm
        matches = []
        
        return matches
