requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0

# Machine Learning
tensorflow>=2.13.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0

# Database
psycopg2-binary>=2.9.0
alembic>=1.11.0

# Authentication
pyjwt>=2.8.0
bcrypt>=4.0.0

# API & Utils
requests>=2.31.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
